Add Pull Request Template
Added a new PR template with sections for specifying the type of PR
(Refactor, Feature, Bug Fix, Optimization, Documentation Update),
a concise description of changes, related tickets, resources for
reviewers, and a checkbox for tests with a target of 90% code coverage.
Also included a field to specify the target branch for merging.